move version info to qulab
