+++ lab/__init__.py
@@ -8,4 +8,4 @@ from ._bootstrap import (connect_db, get_current_notebook, get_current_user,
 from ._plot import imshow, make_figure_for_app, make_figures_for_App, plot
 from ._query import query
 from ._unit import unit
-from ._version import __version__
+from qulab._version import __version__
